+++ packages/x-adapter-platform/src/mappers/responses/ai/tasks-response.mapper.ts
@@ -1,0 +1,13 @@
+import type { AiTasksResponse } from '@empathyco/x-types'
+import type { PlatformAiTasksResponse } from '../../../types'
+import { schemaMapperFactory } from '@empathyco/x-adapter'
+import { aiTasksResponseSchema } from '../../../schemas'
+
+/**
+ * Default implementation for the AiTasksResponseMapper.
+ *
+ * @public
+ */
+export const aiTasksResponseMapper = schemaMapperFactory<PlatformAiTasksResponse, AiTasksResponse>(
+  aiTasksResponseSchema,
+)