+++ packages/x-adapter-platform/src/schemas/requests/ai/tasks-request.schema.ts
@@ -1,0 +1,7 @@
+import type { AiTasksRequest } from '@empathyco/x-types'
+import type { PlatformAiTasksRequest } from '../../../types'
+import { createMutableSchema } from '@empathyco/x-adapter'
+
+export const aiTasksRequestSchema = createMutableSchema<AiTasksRequest, PlatformAiTasksRequest>({
+  taskId: 'taskId',
+})