+++ packages/x-adapter-platform/src/types/responses/ai/questions-response.model.ts
@@ -1,0 +1,3 @@
+import type { AiQuestionsResponse } from '@empathyco/x-types'
+
+export interface PlatformAiQuestionsResponse extends AiQuestionsResponse {}