+++ packages/x-adapter-platform/src/types/responses/index.ts
@@ -1,3 +1,4 @@
+export * from './ai'
 export * from './experience-controls-response.model'
 export * from './identifier-results-response.model'
 export * from './next-queries-response.model'