+++ packages/x-types/src/request/ai/index.ts
@@ -1,0 +1,2 @@
+export * from './questions.model'
+export * from './tasks.model'