+++ packages/x-types/src/x-components-adapter.model.ts
@@ -1,5 +1,7 @@
 import type { EndpointAdapter } from '@empathyco/x-adapter'
 import type {
+  AiQuestionsRequest,
+  AiTasksRequest,
   ExperienceControlsRequest,
   IdentifierResultsRequest,
   NextQueriesRequest,
@@ -13,6 +15,8 @@ import type {
   TaggingRequest,
 } from './request'
 import type {
+  AiQuestionsResponse,
+  AiTasksResponse,
   ExperienceControlsResponse,
   IdentifierResultsResponse,
   NextQueriesResponse,
@@ -41,4 +45,6 @@ export interface XComponentsAdapter {
   tagging: EndpointAdapter<TaggingRequest, void>
   semanticQueries: EndpointAdapter<SemanticQueriesRequest, SemanticQueriesResponse>
   experienceControls: EndpointAdapter<ExperienceControlsRequest, ExperienceControlsResponse>
+  aiQuestions: EndpointAdapter<AiQuestionsRequest, AiQuestionsResponse>
+  aiTasks: EndpointAdapter<AiTasksRequest, AiTasksResponse>
 }